fix(core): align Hotkey action contract
