+++ packages/core/src/core/ui/hotkey/core.ts
@@ -1,14 +1,20 @@
-import type { InputAction } from '../input-action/input-action';
+export type HotkeyActionName =
+  | 'togglePaused'
+  | 'toggleMuted'
+  | 'toggleFullscreen'
+  | 'toggleSubtitles'
+  | 'togglePictureInPicture'
+  | 'seekStep'
+  | 'volumeStep'
+  | 'speedUp'
+  | 'speedDown'
+  | 'seekToPercent';
 
 export interface HotkeyProps {
   /** Key pattern to match, such as `Space`, `ArrowRight`, or `Mod+k`. */
   keys: string;
-  /**
-   * Player action to run when the key pattern matches. Built-in actions are `togglePaused`, `toggleMuted`,
-   * `toggleFullscreen`, `toggleSubtitles`, `togglePictureInPicture`, `seekStep`, `seekToPercent`, `volumeStep`,
-   * `speedUp`, and `speedDown`.
-   */
-  action: InputAction;
+  /** Player action to run when the key pattern matches. */
+  action: HotkeyActionName;
   /** Numeric value passed to actions such as `seekStep`, `volumeStep`, and `seekToPercent`. */
   value?: number | undefined;
   /** Whether the hotkey is disabled. */
