+++ site/src/content/docs/reference/hotkey.mdx
@@ -6,6 +6,7 @@ description: Register a keyboard shortcut that invokes a player action
 ---
 
 import ComponentReference from "@/components/docs/api-reference/ComponentReference.astro";
+import DocsLink from "@/components/docs/DocsLink.astro";
 import FrameworkCase from "@/components/docs/FrameworkCase.astro";
 import StyleCase from "@/components/docs/StyleCase.astro";
 import Demo from "@/components/docs/demos/Demo.astro";
@@ -37,7 +38,7 @@ import basicUsageHtmlTs from "@/components/docs/demos/hotkey/html/css/BasicUsage
 
 ## Behavior
 
-`Hotkey` is a behavior component: it renders no visible interface. Place it inside a `Container` to register one key pattern and player action. By default, the shortcut responds while the player container has focus. Set `target="document"` only when the shortcut should apply across the entire document.
+`Hotkey` is a behavior component: it renders no visible interface. Place it inside a <DocsLink slug="reference/player-container">Container</DocsLink> to register one key pattern and player action. By default, the shortcut responds while the player container has focus. Set `target="document"` only when the shortcut should apply across the entire document.
 
 Key patterns are case-insensitive and may use:
 
@@ -46,20 +47,7 @@ Key patterns are case-insensitive and may use:
 - `Mod` for `Meta` on macOS and `Ctrl` elsewhere.
 - The special `0-9` range for digit-based seeking.
 
-The following player actions are built in:
-
-| Action | Behavior | `value` |
-| --- | --- | --- |
-| `togglePaused` | Toggle play and pause | Not used |
-| `toggleMuted` | Toggle mute | Not used |
-| `toggleFullscreen` | Enter or exit fullscreen | Not used |
-| `toggleSubtitles` | Show or hide subtitles | Not used |
-| `togglePictureInPicture` | Enter or exit picture-in-picture | Not used |
-| `seekStep` | Seek relative to the current time | Seconds to seek; use a negative number to seek backward |
-| `volumeStep` | Adjust the current volume | Volume delta, such as `0.05` or `-0.05` |
-| `speedUp` | Select the next configured playback rate | Not used |
-| `speedDown` | Select the previous configured playback rate | Not used |
-| `seekToPercent` | Seek to a percentage of the duration | Percentage from 0 to 100; when omitted with `keys="0-9"`, the pressed digit supplies it |
+Choose an `action` from the API reference below. Pass `value` as seconds for `seekStep`, a volume delta such as `0.05` or `-0.05` for `volumeStep`, or a percentage from 0 to 100 for `seekToPercent`. When `value` is omitted from `seekToPercent` with `keys="0-9"`, the pressed digit supplies the percentage.
 
 Toggle actions ignore held-key repeats. Step, rate, and percentage actions may repeat. A matched shortcut prevents the browser's default behavior. Unmodified shortcuts are ignored in editable fields, and <kbd>Space</kbd> or <kbd>Enter</kbd> still activates focused buttons, links, and sliders normally.
